Don't replace DatadogAgentlessAPICall on update (#196)
If we change the `PhysicalResourceId` on update, the resource will be
replaced. This means that in the cleanup phase the Lambda will be
deleted, and all the agentless scan options will end up disabled.

Using a stable `PhysicalResourceId` such as the function ARN prevents
this problem.

Note that the Lambda still doesn't support update requests, so changes
to the scan options parameters will not be reflected. This is a lesser
problem that we can fix later.