Send role ARNs in Agentless API call (#232)
We cannot start scheduling agentless scans until the roles are crawled
by the AWS integration. These crawls are infrequent, due to AWS rate
limits.

By sending the ARNs of the roles in the activation request, we can
request an immediate crawl of these specific roles, greatly decreasing
the time it takes to schedule the first scan.