Do not fail stack deletion on API call errors
For most client errors, the best option is to continue with the stack
deletion, since users have no way to fix the request, and at least this
way they can clean up the scanner resources.

An exception might be 429 Too Many Requests, where one could just try
deleting the stack later, but it really doesn't make sense to block a
stack deletion for that reason.