agentless: less strict checking of response status codes (#243)
Checking specific values of status codes causes more harm than good.
Throwing errors when we get a 200 instead of 204 has no benefit. Either
way, the request succeeded. Throwing an error that users have no way
to handle just causes unnecessary stack failures.

Even unsuccessful status codes are not necessarily a problem. If a
delete request fails with 404, that just means that the resource has
already been deleted, i.e. the system is already in the desired state.
It makes no sense to fail the stack deletion because of this, creating
a problem where there was none.

For most client errors, the best option is to continue with the stack
deletion, since users have no way to fix the request, and at least this
way they can clean up the scanner resources.

An exception might be 429 Too Many Requests, where one could just try
deleting the stack later, but it really doesn't make sense to block a
stack deletion for that reason.