Send AgentlessAPICall Lambda ARN to backend (#255)
Sometimes there are delays in crawling resources after installing the
integration. This makes it impossible for us to validate that the
scanner deployment is working properly.

By sending this Lambda ARN to the backend, we can scan it immediately.

Although the DatadogAgentlessAPICall is not a very interesting resource
to scan since it does not have dependencies, it is enough to validate
that the scanner is working properly.